Fix condition to raise ValueError if module app.py does not exist

+++ tests/test_expected_output.py
@@ -19,6 +19,7 @@ APP_NAMES = [p.name for p in FIXTURES_PATH.iterdir() if p.is_dir() and p != DIRS
 @pytest.mark.parametrize('app_name', APP_NAMES)
 def test_output(tmp_path, monkeypatch, app_name, module_name='app'):
     app_path = FIXTURES_PATH / app_name
+    module_path = app_path / str(module_name + '.py')
 
     # Add FIXTURES_PATH to sys.path, the list of directories that `import`
     # looks in
@@ -30,16 +31,21 @@ def test_output(tmp_path, monkeypatch, app_name, module_name='app'):
     expected = app_path / 'test_expected_output'
     print("expected_path:", expected)
 
-    freeze(app, tmp_path)
 
-    if not expected.exists():
-        if 'TEST_CREATE_EXPECTED_OUTPUT' in os.environ:
-            shutil.copytree(tmp_path, expected)
-        else:
-            raise AssertionError(
-                f'Expected output directory ({expected}) does not exist. '
-                + f'Run with TEST_CREATE_EXPECTED_OUTPUT=1 to create it'
-            )
+    if not module_path.exists():
+        with pytest.raises(ValueError):
+            freeze(app, tmp_path)
+    else:
+        freeze(app, tmp_path)
+
+        if not expected.exists():
+            if 'TEST_CREATE_EXPECTED_OUTPUT' in os.environ:
+                shutil.copytree(tmp_path, expected)
+            else:
+                raise AssertionError(
+                    f'Expected output directory ({expected}) does not exist. '
+                    + f'Run with TEST_CREATE_EXPECTED_OUTPUT=1 to create it'
+                )
 
     assert_dirs_same(tmp_path, expected)
